requirements.txt
python-dotenv==1.0.0
aws-cdk-lib==2.181.1
constructs>=10.0.0,<11.0.0
jinja2==3.1.6
